port-ext-tex: changed texnum data size, properly allowing texture names to have more than 4 digits

+++ port/include/ext_tex.h
@@ -5,8 +5,8 @@
 
 s32 extTexInit();
 void extTexFree();
-u8 *extTexLoad(u8 type, u16 id, u16 texnum, u32 *width, u32 *height);
-u8 extTexExists(u8 type, u16 id, u16 texnum);
+u8 *extTexLoad(u8 type, u16 id, s32 texnum, u32 *width, u32 *height);
+u8 extTexExists(u8 type, u16 id, s32 texnum);
 u8 extTexFontID(struct font *font);
 
 #endif
